Consolidate error cases into one file since the tests weren't breaking each other
